Revert "GitHub Action: Fix authorized-changes-detection (#1543)"
This reverts commit 6a910fc1721227f4f9ad34caedf7b0931dc67f45.
