+++ .github/workflows/authorized-changes-detection.yml
@@ -1,0 +1,39 @@
+##############################################################################
+##############################################################################
+#
+# NOTE!
+#
+# Please read the README.md file in this directory that defines what should 
+# be placed in this file
+#
+##############################################################################
+##############################################################################
+
+name: Checking workflow files
+on:
+  pull_request:
+    paths:
+      - '.github/**'   
+      - 'env.example'
+      - '.node-version'
+      - '.husky/**'
+      - 'scripts/**'
+      - 'package.json'
+      - 'tsconfig.json'
+      - '.gitignore'
+      - '.eslintrc.json'
+      - '.eslintignore '
+      - 'vite.config.ts'
+      - 'docker-compose.yaml'
+      - 'Dockerfile'      
+      - 'CODEOWNERS'
+      - 'LICENSE'
+      - 'setup.ts'
+      
+jobs:
+  Checking-for-unauthorized-file-changes:
+    name: Checking for unauthorized file changes
+    runs-on: ubuntu-latest
+    steps:
+      - name: Unauthorized file modification in PR
+        run: exit 1
